requirements.txt
amqp==2.4.2
astroid==2.2.5
beautifulsoup4==4.7.1
billiard==3.6.0.0
boto3==1.9.145
botocore==1.12.145
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
coverage==4.5.3
dj-database-url==0.5.0
Django==2.2.1
django-celery-beat==1.5.0
django-celery-results==1.1.0
django-cors-headers==2.5.3
django-debug-toolbar==1.11
# django-heroku==0.3.1
django-json-widget==0.2.0
django-model-utils==3.1.2
django-money==0.14.4
django-storages==1.7.1
django-timezone-field==3.0
djangorestframework==3.9.3
docutils==0.14
entrypoints==0.3
flake8==3.7.7
googlemaps==3.0.2
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
isort==4.3.18
jmespath==0.9.4
kombu==4.5.0
lazy-object-proxy==1.4.0
lxml==4.3.3
mccabe==0.6.1
numpy==1.16.3
Pillow==6.0.0
# psycopg2==2.8.2
py-moneyed==0.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pygeocoder==1.2.5
pylint==2.3.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
python-crontab==2.3.6
python-dateutil==2.8.0
pytz==2019.1
redis==3.2.1
requests==2.21.0
s3transfer==0.2.0
scipy==1.2.1
# selenium==3.141.0
serpy==0.3.1
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
typed-ast==1.3.5
urllib3==1.24.3
vine==1.3.0
webcolors==1.8.1
webencodings==0.5.1
whitenoise==4.1.2
wrapt==1.11.1
